Transactions refactor
