+++ backend/routes/index.ts
@@ -1,8 +1,13 @@
 var express = require('express');
 var router = express.Router();
-var cors = require('cors')
-var plaid = require('plaid');``
-var moment = require('moment');
+var cors = require('cors');
+var plaid = require('plaid');
+
+const plaidClient = new plaid.Client({
+    clientID: process.env.PLAID_CLIENT_ID,
+    secret: process.env.PLAID_SECRET,
+    env: process.env.NODE_ENV.toUpperCase() === 'PRODUCTION' ? plaid.environments.production : plaid.environments.sandbox
+});
 
 router.use(cors());
 
@@ -10,12 +15,6 @@ import * as _ from 'lodash';
 import checkJwt from '../lib/middleware/secured';
 import * as queries from '../queries';
 import { changeUserPassword, getManagementToken, sendResetPasswordLink, decodeIDToken } from '../lib/middleware/userInfo';
-
-const plaidClient = new plaid.Client({
-  clientID: process.env.PLAID_CLIENT_ID,
-  secret: process.env.PLAID_SECRET,
-  env: process.env.NODE_ENV.toUpperCase() === 'PRODUCTION' ? plaid.environments.production : plaid.environments.sandbox
-});
 
 let managementToken = '';
 
@@ -48,123 +47,6 @@ router.post('/access-token', decodeIDToken, checkJwt, async function (req, res, 
   res.json({});
 });
 
-router.get('/transactions', decodeIDToken, checkJwt, async function (req, res, next) {
-  try{
-    // 1. Exchange public token for access token
-    const email = req.body.user_email;
-    const getBusinessResponse: any = await queries.getBusinessByEmail(email);
-
-    let hours = 0;
-    if (getBusinessResponse.plaidLastPull) {
-      hours = moment().diff(getBusinessResponse.plaidLastPull, 'hours');
-    } else {
-      queries.setPlaidLastPull(getBusinessResponse.businessID);
-    }
-
-    const businessLocationsForBusiness = await queries.getBusinessLocation(email);
-    const defaultBusinessLocationId = businessLocationsForBusiness.businessLocationID; // Temporarily default to user's first business location
-
-    if (hours >= 24) {
-      const accessToken = getBusinessResponse.plaidAccessToken;
-
-      // 2. Fetch transactions from Plaid
-      const now = moment();
-      const today = now.format('YYYY-MM-DD');
-      const startOfYear = moment().startOf('year').format('YYYY-MM-DD');
-
-      let options = {
-        count: 100,
-        offset: 0
-      };
-
-      let actualCount = options.offset;
-      let totalTransactions = options.count; //Just to enter the loop
-      let newTransactions = [];
-      let promises = [];
-
-      while(actualCount < totalTransactions) {
-        let transactionsResponse = await plaidClient.getTransactions(accessToken, startOfYear, today, options);
-        totalTransactions = transactionsResponse.total_transactions;
-
-        actualCount += options.count;
-        options.offset = actualCount;
-
-        // 3. Save transactions to database
-        const transactions = await queries.getTransactions(defaultBusinessLocationId);
-
-
-        for (const transaction of transactionsResponse.transactions) {
-          const foundTransaction = _.find(transactions, { 'amount': transaction.amount, 'date': transaction.date, 'name': transaction.name });
-
-          if (!foundTransaction) {
-            let promise = new Promise( resolve => {
-              let categoryId;
-              if (transaction.amount < 0) {
-                categoryId = 43;
-                newTransactions.push({
-                  businessLocationID: defaultBusinessLocationId,
-                  transactionName: transaction.name,
-                  categoryID: categoryId,
-                  amount: transaction.amount,
-                  date: transaction.date,
-                  address: transaction.location.address,
-                  city: transaction.location.city,
-                  region:  transaction.location.region,
-                  rawData: JSON.stringify(transaction)
-                });
-                resolve();
-              } else {
-                queries.getCategoryForTransaction(transaction.name, defaultBusinessLocationId, businessLocationsForBusiness.vertical, Number(transaction.category_id))
-                .then(categoryId => {
-                  newTransactions.push({
-                    businessLocationID: defaultBusinessLocationId,
-                    transactionName: transaction.name,
-                    categoryID: categoryId,
-                    amount: transaction.amount,
-                    date: transaction.date,
-                    address: transaction.location.address,
-                    city: transaction.location.city,
-                    region:  transaction.location.region,
-                    rawData: JSON.stringify(transaction)
-                  });
-                  resolve();
-                });
-              }
-            });
-            promises.push(promise);
-          }
-        }
-      }
-
-      // Insert transaction into database
-      if (newTransactions.length) {
-        await Promise.all(promises);
-        await queries.saveTransactions(newTransactions);
-      }
-    }
-
-    
-
-    // 4. Return transactions in response
-    let latestTransactions = await queries.getTransactions(defaultBusinessLocationId);
-    latestTransactions = latestTransactions.map(transaction => {
-      return {
-        type: transaction.type.toLowerCase(),
-        amount: transaction.amount,
-        name: transaction.name,
-        category: transaction.account,
-        categoryId: transaction.categoryID
-      };
-    });
-
-    res.json(latestTransactions);
-  } catch(error) {
-    console.log(error);
-
-    res.json(error);
-  }
-});
-
 router.get('/expense-category', decodeIDToken, checkJwt, async function (req, res, next) {
   try{
     const categoryId = req.query.categoryId;
@@ -183,39 +65,6 @@ router.get('/expense-category', decodeIDToken, checkJwt, async function (req, re
         category: transaction.account,
         categoryId: transaction.categoryID,
         transactionId: transaction.transactionID,
-        date: transaction.date
-      };
-    });
-
-    res.json(latestTransactions);
-  } catch(error) {
-    console.log(error);
-
-    res.json(error);
-  }
-});
-
-router.get('/transaction', decodeIDToken, checkJwt, async function (req, res, next) {
-  try {
-    const transactionId = req.query.transactionId;
-    const email = req.body.user_email;
-    const businessLocationsForBusiness = await queries.getBusinessLocation(email);
-    const defaultBusinessLocationId = businessLocationsForBusiness.businessLocationID;
-
-    let latestTransactions = await queries.getTransactions(defaultBusinessLocationId);
-    latestTransactions = latestTransactions.filter(transaction => transaction.transactionID === Number(transactionId));
-
-    latestTransactions = latestTransactions.map(transaction => {
-      return {
-        type: transaction.type.toLowerCase(),
-        amount: transaction.amount,
-        name: transaction.name,
-        category: transaction.account,
-        categoryId: transaction.categoryID,
-        transactionId: transaction.transactionID,
-        address: transaction.address,
-        city: transaction.city,
-        region:  transaction.region,
         date: transaction.date
       };
     });
@@ -329,14 +178,6 @@ router.post('/reset-password', async function (req, res, next) {
 
 });
 
-router.get('/transaction-categories', decodeIDToken, checkJwt, async function (req, res, next) {
-
-  const response = await queries.getChartOfAccountsCategories(req.query.vertical);
-
-  res.json(response);
-
-});
-
 router.get('/location', decodeIDToken, checkJwt, async function (req, res, next) {
 
   const email = req.body.user_email;
@@ -345,24 +186,6 @@ router.get('/location', decodeIDToken, checkJwt, async function (req, res, next)
 
   res.json(response);
 
-});
-
-router.put('/transactions', decodeIDToken, checkJwt, async function (req, res, next) {
-  try {
-    const transaction = req.body.transaction;
-    const email = req.body.user_email;
-    const businessLocationsForBusiness = await queries.getBusinessLocation(email);
-    const defaultBusinessLocationId = businessLocationsForBusiness.businessLocationID;
-    const response = await queries.updateTransaction(transaction);
-
-    queries.updateSimilarTransactionsForUser(transaction.name, transaction.categoryId, defaultBusinessLocationId);
-
-    res.json(response);
-  } catch(error) {
-    console.log(error);
-
-    res.json(error);
-  }
 });
 
 router.get('/has-plaid-token', decodeIDToken, checkJwt, async function (req, res, next) {
@@ -380,5 +203,6 @@ router.get('/has-plaid-token', decodeIDToken, checkJwt, async function (req, res
 });
 
 router.use('/taxes', require('./taxes'));
+router.use('/transactions', require('./transactions'));
 
 module.exports = router;
